requirements.txt
pandas==2.2.3 
numpy==2.1.2
scikit-learn==1.5.2
statsmodels==0.14.4
matplotlib==3.9.2
pandas_datareader-0.10.0
yfinance-0.2.48
seaborn-0.13.2
